force UTC when comparing time
Due to the hacky method localtime in TimeWithZonePatch, the default time
comparison doesn't work when

DateAndTime::Compatibility.preserve_timezone = true

and we compare utc time vs time with dynamic time zone.
As a result, we must force both side of comparison to utc in order to do
a fair comparison.